pyproject.toml
[project]
name = "dbt-unified-mcp"
version = "0.1.0"
description = "Unified multi-client dbt Cloud MCP server for RKD"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "boto3>=1.34",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "black>=24.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.black]
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
